InboundTelemetryService/requirements.txt
flask
kafka-python
requests